separating stuff
