+++ Makefile
@@ -8,7 +8,7 @@ HEADERS := -I ./inc -I $(LIBMLX)/include -I $(LIBFT)
 LIBS    := $(LIBMLX)/build/libmlx42.a $(LIBFT)/libft.a -ldl -lglfw -pthread -lm
 SRCS    := src/fill_map.c src/get_next_line.c src/parse.c src/pre_map_parse.c src/map_parse.c \
            src/utils_parse.c src/utils_parse_2.c src/parse_main.c src/render.c src/input.c src/raycast.c \
-           src/movements.c src/assets.c src/textures.c
+           src/movements.c src/assets.c src/textures.c src/init_cub.c
 
 OBJDIR  := .obj
 OBJS    := $(SRCS:%.c=$(OBJDIR)/%.o)
